feat: relax dependency on Momento SDK
While investigating this issue:

https://github.com/momentohq/momento-node-ioredis-client/issues/18

I found that given our current example code it's fairly easy for
a user to get into a situation where they end up with two different
versions of our SDK dependency, and they get weird type errors and
can't compile.

The primary reason for this is that you need to construct a CacheClient
from the SDK in order to pass to the constructor for the Redis client
from this library. If you have typical eslint rules enabled, you
will be warned that you can't reference classes from the SDK such as
`CacheClient` without having an explicit dependency on the SDK in
your package.json. If, in attempting to resolve that problem, the user
does `npm install @gomomento/sdk`, and the version that gets installed
is not the exact same version that the momento-ioredis-client is pinned
to, then they will get build errors.

I think we should do two things to remedy this:
1. relax this dependency version; I normally advocate for strict dependency
  versions but when you have to pass objects across API boundaries it
  becomes problematic.
2. Re-export the CacheClient, Configuration, and any other objects from the
  sdk that users of this momento-ioredis library will need to pass to the redis
  client constructor; that way they can avoid having to add an explicit
  dependency on the momento sdk at all.

This commit does the first of those 2 items. In a subsequent PR we'll address
the second one.